+++ pocket/Makefile
@@ -1,7 +1,7 @@
 GOOS := $(if $(GOOS),$(GOOS),linux)
 GOARCH := $(if $(GOARCH),$(GOARCH),amd64)
-GO=GO15VENDOREXPERIMENT="1" CGO_ENABLED=0 GOOS=$(GOOS) GOARCH=$(GOARCH) GO111MODULE=off go
-GOTEST=GO15VENDOREXPERIMENT="1" CGO_ENABLED=1 GO111MODULE=off go test # go race detector requires cgo
+GO=GO15VENDOREXPERIMENT="1" CGO_ENABLED=0 GOOS=$(GOOS) GOARCH=$(GOARCH) GO111MODULE=on go
+GOTEST=GO15VENDOREXPERIMENT="1" CGO_ENABLED=1 GO111MODULE=on go test # go race detector requires cgo
 VERSION   := $(if $(VERSION),$(VERSION),latest)
 
 PACKAGES := $$(go list ./...| grep -vE 'vendor' )
@@ -11,8 +11,8 @@ GOFILTER  := grep -vE 'vendor|render.Delims|bindata_assetfs|testutil|\.pb\.go'
 GOCHECKER := $(GOFILTER) | awk '{ print } END { if (NR > 0) { exit 1 } }'
 GOLINT    := go list ./... | grep -vE 'vendor' | xargs -L1 -I {} golint {} 2>&1 | $(GOCHECKER)
 
-LDFLAGS += -X "github.com/pingcap/abclient/util.BuildTS=$(shell date -u '+%Y-%m-%d %I:%M:%S')"
-LDFLAGS += -X "github.com/pingcap/abclient/util.BuildHash=$(shell git rev-parse HEAD)"
+LDFLAGS += -X "github.com/pingcap/pocket/util.BuildTS=$(shell date -u '+%Y-%m-%d %I:%M:%S')"
+LDFLAGS += -X "github.com/pingcap/pocket/util.BuildHash=$(shell git rev-parse HEAD)"
 
 GOBUILD=$(GO) build -ldflags '$(LDFLAGS)'
 
@@ -21,7 +21,7 @@ default: build
 build: client
 
 client:
-	$(GOBUILD) $(GOMOD) -o bin/abclient *.go
+	$(GOBUILD) $(GOMOD) -o bin/pocket *.go
 
 clean:
 	@rm -rf bin/*
