+++ pocket/connection/types.go
