+++ pocket/core/handler.go
@@ -5,8 +5,8 @@ import (
 	"math/rand"
 	"github.com/juju/errors"
 	"github.com/ngaut/log"
-	"github.com/pingcap/tipocket/abclient/executor"
-	"github.com/pingcap/tipocket/abclient/pkg/types"
+	"github.com/pingcap/tipocket/pocket/executor"
+	"github.com/pingcap/tipocket/pocket/pkg/types"
 )
 
 func (e *Executor) startHandler() {
