+++ pocket/core/reproduce.go
@@ -12,8 +12,8 @@ import (
 	"sort"
 	"github.com/juju/errors"
 	"github.com/ngaut/log"
-	"github.com/pingcap/tipocket/abclient/util"
-	"github.com/pingcap/tipocket/abclient/pkg/types"
+	"github.com/pingcap/tipocket/pocket/util"
+	"github.com/pingcap/tipocket/pocket/pkg/types"
 )
 
 var (
