+++ pocket/executor/common.go
