+++ pocket/executor/executor.go
@@ -6,9 +6,9 @@ import (
 	"os"
 	"regexp"
 	"sync"
-	"github.com/pingcap/tipocket/abclient/connection"
-	"github.com/pingcap/tipocket/abclient/pkg/logger"
-	"github.com/pingcap/tipocket/abclient/pkg/types"
+	"github.com/pingcap/tipocket/pocket/connection"
+	"github.com/pingcap/tipocket/pocket/pkg/logger"
+	"github.com/pingcap/tipocket/pocket/pkg/types"
 	"github.com/juju/errors"
 	"github.com/ngaut/log"
 	smith "github.com/pingcap/tipocket/go-sqlsmith"
