+++ pocket/executor/misc.go
