+++ pocket/executor/reproduce.go
@@ -7,8 +7,8 @@ import (
 	"strconv"
 	"strings"
 	"github.com/ngaut/log"
-	"github.com/pingcap/tipocket/abclient/util"
-	"github.com/pingcap/tipocket/abclient/pkg/types"
+	"github.com/pingcap/tipocket/pocket/util"
+	"github.com/pingcap/tipocket/pocket/pkg/types"
 )
 
 var (
