+++ pocket/test-binlog-re.sh
